hsy商户

+++ service/hsyuser/user-api/src/main/java/com/jkm/hsy/user/service/HsyMerchantAuditService.java
@@ -53,7 +53,7 @@ public interface HsyMerchantAuditService {
     void updateAccount(Long accountID, Long id);
 
 
-    void stepChange(Long uid);
+    void stepChange(int uid);
 
-    Long getUid(Long id);
+    int getUid(Long id);
 }
